rename file with REST API helper functions
