+++ cfme/tests/configure/test_rest_access_control.py
@@ -6,10 +6,10 @@ from cfme import Credential
 from cfme import test_requirements
 from cfme.configure.access_control import User
 from cfme.login import login, login_admin
-from cfme.rest import groups as _groups
-from cfme.rest import roles as _roles
-from cfme.rest import tenants as _tenants
-from cfme.rest import users as _users
+from cfme.rest.gen_data import groups as _groups
+from cfme.rest.gen_data import roles as _roles
+from cfme.rest.gen_data import tenants as _tenants
+from cfme.rest.gen_data import users as _users
 from utils.wait import wait_for
 from utils import error
