+++ cfme/tests/infrastructure/test_rest_providers.py
@@ -3,7 +3,7 @@ import pytest
 
 from cfme import test_requirements
 from cfme.configure.configuration import server_roles_disabled
-from cfme.rest import a_provider as _a_provider
+from cfme.rest.gen_data import a_provider as _a_provider
 from utils.virtual_machines import deploy_template
 from utils.version import current_version
 from utils.wait import wait_for
