+++ cfme/tests/infrastructure/test_rest_templates.py
@@ -2,9 +2,9 @@
 import pytest
 
 from cfme import test_requirements
-from cfme.rest import a_provider as _a_provider
-from cfme.rest import mark_vm_as_template
-from cfme.rest import vm as _vm
+from cfme.rest.gen_data import a_provider as _a_provider
+from cfme.rest.gen_data import mark_vm_as_template
+from cfme.rest.gen_data import vm as _vm
 from utils import error
 
 pytestmark = [pytest.mark.ignore_stream("5.4"), test_requirements.rest]
