+++ cfme/tests/infrastructure/test_timelines.py
@@ -4,8 +4,8 @@ import pytest
 
 from cfme.common.vm import VM
 from cfme.infrastructure.provider import InfraProvider
-from cfme.rest import a_provider as _a_provider
-from cfme.rest import vm as _vm
+from cfme.rest.gen_data import a_provider as _a_provider
+from cfme.rest.gen_data import vm as _vm
 from cfme.web_ui import InfoBlock, toolbar, jstimelines
 from cfme.exceptions import ToolbarOptionGreyedOrUnavailable
 from utils import testgen
