+++ cfme/tests/infrastructure/test_vm_rest.py
@@ -2,8 +2,8 @@
 import pytest
 
 from cfme import test_requirements
-from cfme.rest import a_provider as _a_provider
-from cfme.rest import vm as _vm
+from cfme.rest.gen_data import a_provider as _a_provider
+from cfme.rest.gen_data import vm as _vm
 from utils.wait import wait_for
