+++ cfme/tests/intelligence/test_chargeback.py
@@ -6,7 +6,7 @@ import cfme.web_ui.flash as flash
 import utils.error as error
 
 from cfme import test_requirements
-from cfme.rest import rates as _rates
+from cfme.rest.gen_data import rates as _rates
 from utils.update import update
 from utils.wait import wait_for
