+++ cfme/tests/services/test_generic_service_catalogs.py
@@ -3,7 +3,7 @@ import fauxfactory
 import pytest
 
 from cfme.automate.service_dialogs import ServiceDialog
-from cfme.rest import service_catalogs as _service_catalogs
+from cfme.rest.gen_data import service_catalogs as _service_catalogs
 from cfme.services.catalogs.catalog_item import CatalogItem
 from cfme.services.catalogs.service_catalogs import ServiceCatalogs
 from cfme.services.catalogs.catalog_item import CatalogBundle
