+++ cfme/tests/test_rest.py
@@ -5,8 +5,8 @@ import fauxfactory
 import utils.error as error
 
 from cfme import test_requirements
-from cfme.rest import vm as _vm
-from cfme.rest import arbitration_settings
+from cfme.rest.gen_data import vm as _vm
+from cfme.rest.gen_data import arbitration_settings
 from utils.providers import setup_a_provider as _setup_a_provider
 from utils.version import current_version
 from utils import testgen, version
